requirements.txt
absl-py==1.3.0
asttokens==2.1.0
backcall==0.2.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
gensim==4.2.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
preshed==3.0.8
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-nlp @ git+https://github.com/PetrochukM/PyTorch-NLP.git@d7814a297811c9b0dfb285fe0475098b86f3d348
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
smart-open==5.2.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.5
threadpoolctl==3.1.0
torch==1.13.0
torchtext==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
